Otimizações

+++ src/dynamicArray.cpp
@@ -39,7 +39,7 @@ void DynamicArray<T>::append(T item, int idx)
     }
 
     // Abre espaço para um novo item movendo todos os items que estão a direita do índice, 1 espaço
-    for (int i = currentSize; i > idx; i--)
+    for (size_t i = currentSize; i > idx; i--)
     {
         arr[i] = arr[i - 1];
     }
